requirements.txt
ops
pyOpenSSL
python-hosts==1.0.1
git+https://github.com/canonical/ops-lib-nrpe#egg=ops-lib-nrpe
git+https://opendev.org/openstack/charm-ops-interface-tls-certificates#egg=interface_tls_certificates
git+https://github.com/phvalguima/ops-coordinator#egg=ops-coordinator
